--- src/main/resources/allure.properties
@@ -1,2 +1,0 @@
-allure.results.directory = target/allure-results
-allure.link.issue.pattern = https://github.com/ita-social-projects/GreenCityTA/issues/{}
